[Bronze V] Title: 사파리월드, Time: 100 ms, Memory: 14052 KB -BaekjoonHub

+++ 백준/Bronze/2420. 사파리월드/사파리월드.java
@@ -8,8 +8,8 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
 
-        long n = Integer.parseInt(st.nextToken());
-        long m = Integer.parseInt(st.nextToken());
+        long n = Long.parseLong(st.nextToken());
+        long m = Long.parseLong(st.nextToken());
 
         System.out.println(Math.abs(n - m));
     }
